unittest: seperate target `longunittest' and `shortunittest'
There are some unittests which took a quite long time, e.g. Test6850611 takes
about 16seconds on my machine.  We want them to shift into a seperate mx target.

In order to move a testcase to `longunittest' you have to replace all `@Test'
with `@LongTest' in the testclass.

`@Test' belongs to `shortunittest'.  The target `unittest` executes both,
`@Test' and `@LongTest'.

Note: Mixing `@Test' and `@LongTest' in a testclass, causes that each testmethod
is executed for both, `shortunittest' and `longunittest', mx targets.